% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abid_lib}
\alias{abid_lib}
\title{Default library file to be used by \code{ABID} to compared digested MALDI spectra against.}
\format{
A list of annotated MALDIquant mass spectra.
}
\source{
Georg.Tscheuschner@bam.de
}
\usage{
data(abid_lib)
}
\description{
Default library file to be used by \code{ABID} to compared digested MALDI spectra against.
}
\keyword{datasets}
