% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGroupFactor.R
\name{GetGroupFactor}
\alias{GetGroupFactor}
\title{GetGroupFactor.}
\usage{
GetGroupFactor(x, gap)
}
\arguments{
\item{x}{Numeric vector.}

\item{gap}{Numeric gap to breaks x into groups.}
}
\value{
A factor vector that can be used to split x into groups.
}
\description{
\code{GetGroupFactor} will group a numeric vector according to a specified gap.
}
\details{
tbd.
}
\keyword{internal}
