% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readABSpec.R
\name{readABSpec}
\alias{readABSpec}
\title{readABSpec.}
\usage{
readABSpec(file = NULL, meta_data = NULL, limit_mz = NULL, ...)
}
\arguments{
\item{file}{Path to a text file containing a MALDI spectrum.}

\item{meta_data}{If provided will be attached to output. Will be guessed from file header if NULL.}

\item{...}{Further parameters to read.table function.}
}
\value{
A MALDIquant object with metadata.
}
\description{
\code{readABSpec} will read a MALDI file.
}
\details{
{readABSpec} will read a MALDI file and convert it to a MALDIquant object with metadata..
}
\examples{
# tbd

}
\keyword{internal}
