% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_ABID_Lib.R
\name{Setup_ABID_Lib}
\alias{Setup_ABID_Lib}
\title{Setup_ABID_Lib.}
\usage{
Setup_ABID_Lib(
  folder = "C:/Users/jlisec/Documents/Projects/ABID/Libraries/ABIDApp",
  limit_n = NULL,
  limit_mz = NULL,
  depr = 0,
  lib_name = "abid"
)
}
\arguments{
\item{folder}{Path to a text file containing a MALDI spectrum.}

\item{limit_n}{limit_n.}

\item{limit_mz}{limit_mz.}

\item{depr}{depr.}

\item{lib_name}{lib_name.}
}
\value{
A rda file with MALDI spectra and meta data.
}
\description{
\code{Setup_ABID_Lib} will Setup_ABID_Lib.
}
\details{
{Setup_ABID_Lib} will rSetup_ABID_Lib.
}
\examples{
\dontrun{
Setup_ABID_Lib(folder="C:/Users/jlisec/Documents/Projects/ABID/Libraries/ABIDApp", limit_mz=7000)
}

}
\keyword{internal}
