% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subclass_peak_list}
\alias{subclass_peak_list}
\title{Default peak lists used by \code{ABID} to assign antibody sub classes to digested MALDI spectra.}
\format{
A list of length = 9 containing two column data frames and an attribute 'info' attached to it:
}
\source{
Georg.Tscheuschner@bam.de
}
\usage{
data(subclass_peak_list)
}
\description{
Default peak lists used by \code{ABID} to assign antibody sub classes to digested MALDI spectra.
}
\keyword{datasets}
