% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abid_app.R
\name{abid_app}
\alias{abid_app}
\title{abid_app.}
\usage{
abid_app(bam_server = TRUE)
}
\arguments{
\item{bam_server}{Set TRUE to include BAM tracking script and privacy-policy link.}
}
\value{
A \code{shinyApp} object.
}
\description{
\code{abid_app} will start the App.
}
\details{
Evaluation of MALDI measurements of digested antibodies.
}
