% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclass_peak_list.R
\docType{data}
\name{subclass_peak_list}
\alias{subclass_peak_list}
\title{Library file.}
\format{
A list of xx spectra.
}
\source{
Georg.Tscheuschner@bam.de
}
\usage{
subclass_peak_list
}
\description{
Library file.
}
\keyword{datasets}
